week2/requirements.txt
black==23.3.0
fastparquet==2023.4.0
hyperopt==0.2.7
mlflow==2.5.0
pandas==2.0.1
prefect==2.10.8
prefect-aws==0.3.1
scikit_learn==1.3.0
seaborn==0.12.2
xgboost==1.7.5
orjson==3.8.1
pydantic==1.10.0
google-cloud==0.34.0
google-cloud-storage==2.10.0
gcsfs
psycopg2-binary
numpy